add BlockGenerator
