+++ road_network_generator.py
@@ -7,75 +7,89 @@ import ghpythonlib.components as ghcomp
 import utils
 import importlib
 import scriptcontext as sc
+import units
 
+from units import Node, Edge, Road, Junction, RoadNetwork
+
+importlib.reload(units)
 importlib.reload(utils)
 TOL = 5
 
 
-class Edge:
-    def __init__(self, curve: geo.Curve, width: float):
-        self.curve = curve  # type: geo.Curve
-        self.width = width  # type: float
-
-    def is_at(self, node: "Node") -> bool:
-        return (
-            self.curve.PointAtStart == node.point or self.curve.PointAtEnd == node.point
-        )
-
-
-class Node:
-    def __init__(self, point: geo.Point3d):
-        self.point = point
-        self.edges = []  # type: List[Edge]
-
-    def add_edges(self, edges: List[Edge]) -> None:
-        self.edges.extend(edges)
-
-    def _key_from_point(self, p: geo.Point3d):
-        """Quantize point to TOL-sized grid so equality and hashing align."""
-        q = TOL
-        return (
-            int(round(p.X / q)),
-            int(round(p.Y / q)),
-            int(round(p.Z / q)),
-        )
-
-    def __eq__(self, other):
-        # Equal if the points fall into the same TOL-sized grid cell
-        if isinstance(other, Node):
-            return self._key_from_point(self.point) == self._key_from_point(other.point)
-        if isinstance(other, geo.Point3d):
-            return self._key_from_point(self.point) == self._key_from_point(other)
-        return NotImplemented
-
-    def __hash__(self):
-        # Hash derived from the same quantized key used in __eq__
-        return hash(self._key_from_point(self.point))
-
-
-class Road:
-    def __init__(self, region: geo.Curve, edge: Edge):
-        self.region = region
-        self.edge = edge
-
-
-class Junction:
-    def __init__(self, region, node: Node):
-        self.region = region
-        self.node = node
-        self.connected_roads = []  # type: List[Road]
-
-
-class RoadNetwork:
+class RoadNetworkGenerator:
     def __init__(self):
-        self.nodes = []  # type: List[Node]
-        self.edges = []  # type: List[Edge]
-        self.roads = []  # type: List[Road]
-        self.junctions = []  # type: List[Junction]
+        pass
 
-    def generate(self, road_data: List[Tuple[geo.Curve, float]]) -> None:
-        self.nodes, self.edges = self._get_node_and_edge(road_data)
-        self.roads, self.junctions = self._get_road_and_junction()
+    def generate(self, road_data: List[Tuple[geo.Curve, float]]) -> RoadNetwork:
+        """도로 네트워크를 생성합니다."""
+
+        road_data = self.snap_roads(road_data, tol=2)
+        nodes, edges = self._get_node_and_edge(road_data)
+        roads, junctions = self._get_road_and_junction(nodes, edges)
+        road_network = RoadNetwork(roads, junctions, nodes, edges)
+        return road_network
+
+    def snap_roads_pt(
+        self, roads: List[Tuple[geo.Curve, float]], tol: float
+    ) -> List[geo.Curve]:
+        """도로 네트워크의 모든 교차점을 스냅합니다."""
+        if len(roads) < 2:
+            return roads
+
+        # 모든 커브 쌍의 교차점 수집
+        all_crvs = [road[0] for road in roads]
+        pts = utils.get_pts_from_crvs(all_crvs)
+        for crv in all_crvs:
+            pts.extend(utils.get_vertices(crv))
+
+        snaped_road_datas = []
+        for road_crv, width in sorted(roads, key=lambda x: x[1]):
+            for pt in pts:
+                if pt.EpsilonEquals(road_crv.PointAtStart, tol):
+                    road_crv = utils.move_curve_endpoint(road_crv, pt, "start")
+
+                if pt.EpsilonEquals(road_crv.PointAtEnd, tol):
+                    road_crv = utils.move_curve_endpoint(road_crv, pt, "end")
+            snaped_road_datas.append((road_crv, width))
+
+        return snaped_road_datas
+
+    def snap_roads_crv(
+        self, roads: List[Tuple[geo.Curve, float]], tol: float
+    ) -> List[geo.Curve]:
+        """도로 네트워크의 모든 교차점을 스냅합니다."""
+        if len(roads) < 2:
+            return roads
+
+        snaped_road_datas = []
+        for road_crv, width in sorted(roads, key=lambda x: x[1]):
+            other_road_crvs = [crv for crv, w in roads if crv != road_crv]
+            for other_road_crv in other_road_crvs:
+                _, param = other_road_crv.ClosestPoint(road_crv.PointAtStart)
+                point_at = other_road_crv.PointAt(param)
+                if point_at.DistanceTo(road_crv.PointAtStart) <= tol:
+                    road_crv = utils.move_curve_endpoint(road_crv, point_at, "start")
+
+                _, param = other_road_crv.ClosestPoint(road_crv.PointAtEnd)
+                point_at = other_road_crv.PointAt(param)
+                if point_at.DistanceTo(road_crv.PointAtEnd) <= tol:
+                    road_crv = utils.move_curve_endpoint(road_crv, point_at, "end")
+
+            snaped_road_datas.append((road_crv, width))
+
+        return snaped_road_datas
+
+    def snap_roads(
+        self, road_data: List[Tuple[geo.Curve, float]], tol: float
+    ) -> List[geo.Curve]:
+        """도로 네트워크의 모든 교차점을 스냅합니다."""
+        if len(road_data) < 2:
+            return road_data
+
+        road_data = self.snap_roads_crv(road_data, tol)
+        road_data = self.snap_roads_pt(road_data, tol)
+
+        return road_data
 
     def _get_node_and_edge(
         self, road_data: List[Tuple[geo.Curve, float]]
@@ -85,14 +99,12 @@ class RoadNetwork:
         curves = [rd[0] for rd in road_data]
         node_pts = utils.get_pts_from_crvs(curves)
         edges = self._get_edges(road_data, node_pts)
-        print("EDGE COUNT:", len(edges))
 
         nodes = []
         for node_pt in node_pts:
             node = Node(node_pt)
             connected_edges = [e for e in edges if e.is_at(node)]
             node.add_edges(connected_edges)
-            print("EDGE COOUNT:", len(connected_edges))
             nodes.append(node)
 
         return nodes, edges
@@ -110,15 +122,17 @@ class RoadNetwork:
                 edges.append(edge)
         return edges
 
-    def _get_road_and_junction(self) -> Tuple[List[Road], List[Junction]]:
+    def _get_road_and_junction(
+        self, nodes: List[Node], edges: List[Edge]
+    ) -> Tuple[List[Road], List[Junction]]:
         junctions = []  # type: List[Junction]
-        for node in self.nodes:
+        for node in nodes:
             junction_region = self._get_junction_region(node)
             junction = Junction(junction_region, node)
             junctions.append(junction)
 
         roads = []  # type: List[Road]
-        for edge in self.edges:
+        for edge in edges:
             road_region = self._get_road_region(edge, junctions)
             road = Road(road_region, edge)
             roads.append(road)
@@ -130,7 +144,6 @@ class RoadNetwork:
         return roads, junctions
 
     def _get_junction_region(self, node: Node) -> geo.Curve:
-
         node_circle = geo.Circle(
             node.point, max(e.width for e in node.edges)
         ).ToNurbsCurve()
@@ -165,7 +178,6 @@ class RoadNetwork:
         if not junction_regions:
             return road_region[0]
 
-        print(road_region, junction_regions)
         diff_region = utils.get_difference_regions(road_region, junction_regions, 0.5)
 
         return max(diff_region, key=lambda r: r.GetLength())
